Move two important methods out of transform/
