+++ osm2streets/src/operations/mod.rs
@@ -1,0 +1,4 @@
+// Particularly important or complex operations on a StreetNetwork
+
+mod collapse_intersection;
+mod collapse_short_road;
